rename publicapi package to wbapi
